add keyboard shortcut of Super+KF
Only triggers if you are in an Elm file

Idea is to match the convention of "Super+KU for uppercase" so you
Super+KF for elm-format.

Ideally it'd be something like Ctrl-EF and Ctrl-EM for format and make,
but I'm not sure how to do that across OSes yet. (since ctrl+e means
go-to-end-of-line for some systems)